pyproject.toml
[tool.poetry]
name = "kronos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
	 { include = "app", from = "" }
	 ]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
fastapi = "^0.110.0"
python-multipart = "^0.0.6"
Jinja2 = "^3.1.2"
pydantic = {extras = ["email"], version = "^2.4.2"}
databases = {extras = ["aiosqlite", "asyncpg"], version = "^0.8.0"}
psycopg = {extras = ["binary"], version = "^3.1.13"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
black = "^23.10.0"
flake8 = "^6.1.0"
autoflake = "^2.2.1"
pytest = "^7.4.2"
anyio = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]